Fix wrong remote IP if server is behind a proxy (#1184)
